made provider and utils.url has been moved to provider.url
