+++ src/exojax/utils/photometry.py
@@ -2,7 +2,7 @@ import numpy as np
 import jax.numpy as jnp
 from exojax.utils.grids import wav2nu
 from exojax.utils.constants import ccgs
-from exojax.utils.url import url_svo_filter
+from exojax.provider.url import url_svo_filter
 from jax.scipy.integrate import trapezoid
